requirements.txt
certifi==2019.6.16
chardet==3.0.4
dateparser==0.7.1
Django==2.2.3
djangorestframework==3.10.2
idna==2.8
psycopg2==2.8.3
Pygments==2.4.2
python-dateutil==2.8.0
python-gedcom==1.0.0
pytz==2019.2
regex==2019.6.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tzlocal==2.0.0
urllib3==1.25.3
